+++ packages/client/editor/src/components/Newsidebar/StyledDivider.tsx
@@ -1,0 +1,14 @@
+import Divider from '@mui/material/Divider'
+import styled from 'styled-components'
+
+const StyledDivider = styled(Divider)(({ theme }) => ({
+  backgroundColor: '#3D454A',
+  borderColor: '#3D454A',
+  height: '1px',
+  marginTop: '4px',
+  marginBottom: '4px',
+  marginLeft: '10px',
+  marginRight: '10px',
+}))
+
+export default StyledDivider
